requirements.txt
fastapi>=0.110,<0.112
uvicorn[standard]>=0.29,<0.31
streamlit>=1.32,<1.40
requests>=2.31,<3.0
httpx>=0.27,<0.29
python-dotenv>=1.0,<2.0

langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0

chromadb==0.4.24
unstructured>=0.12.0,<0.15.0
pandas>=2.0,<3.0

google-generativeai>=0.7.0,<0.9.0